+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 * add a fake transporter that can be used for testing without dependencies on an
   actual connected transporter
+* `log_level`, `transporter`, `heartbeat_interval`, `log_file` and `timeout` are now configurable
+  via environment variable prefixed with `MOLECULER_`
 * updated json to 2.2.0
 * updated oj to 3.7.12
 * updated redis to 4.1.2